+++ engine/core/engine.h
@@ -217,6 +217,7 @@ private:
     std::unique_ptr<SubtitleQueueApplySink>              m_subtitleSink;
     std::unique_ptr<AudioEngineHrtfApplySink>            m_hrtfSink;
     std::unique_ptr<AudioEngineOutputApplySink>          m_outputLayoutSink;  // AX8
+    std::unique_ptr<AudioEngineAirAbsorptionApplySink>   m_airAbsorptionSink; // AX6
     std::unique_ptr<PhotosensitiveStoreApplySink>        m_photosensitiveSink;
     std::unique_ptr<LocalizationServiceApplySink>        m_localizationSink;